remove all non-wallet sigmoji code
