+++ apps/jubmoji-quest/src/lib/localStorage.ts
@@ -37,77 +37,6 @@ export async function loadSigmojiWalletBackup(): Promise<string> {
 }
 
 /**
- * Saves a Sigmoji to the local storage.
- * @param {Sigmoji} sigmoji - The Sigmoji to save.
- */
-export async function saveSigmoji(sigmoji: Sigmoji): Promise<void> {
-  const sigmojis = await loadSigmojis();
-  sigmojis.push(sigmoji);
-  serializeSigmojisInLocalStorage(sigmojis);
-}
-
-/**
- * Updates a Sigmoji in the localStorage.
- * @param {Sigmoji} sigmoji - The Sigmoji to update.
- */
-export async function updateSigmoji(sigmoji: Sigmoji): Promise<void> {
-  const sigmojis = await loadSigmojis();
-  const index = sigmojis.findIndex((s) => s.emojiImg === sigmoji.emojiImg);
-  if (index !== -1) {
-    sigmojis[index] = sigmoji;
-  }
-
-  serializeSigmojisInLocalStorage(sigmojis);
-}
-
-/*
- * Updates the sigmojis in local storage with a new list.
- * Will overwrite any sigmojis that have the same emojiImg with the new list.
- * @param newSigmojis - The new list of sigmojis to update with.
- */
-export async function updateSigmojiList(newSigmojis: Sigmoji[]): Promise<void> {
-  const currentSigmojis = await loadSigmojis();
-  const newSigmojiKeys = newSigmojis.map((s) => s.emojiImg);
-  const updatedSigmojis = newSigmojis.concat(
-    currentSigmojis.filter((s) => !newSigmojiKeys.includes(s.emojiImg))
-  );
-
-  serializeSigmojisInLocalStorage(updatedSigmojis);
-}
-
-/**
- * Saves a user's leaderboard entry to localStorage.
- * @param {object} entry - The leaderboard entry to save.
- * @param {string} entry.pseudonym - The pseudonym of the user.
- * @param {number} entry.score - The score of the user.
- */
-export function saveLeaderboardEntry(entry: {
-  pseudonym: string;
-  score: number;
-}): void {
-  const leaderboardEntries = loadLeaderboardEntries();
-  const serializedEntry = JSON.stringify(entry);
-  if (!leaderboardEntries.includes(serializedEntry)) {
-    leaderboardEntries.push(serializedEntry);
-  }
-
-  window.localStorage["leaderboard"] = JSON.stringify(leaderboardEntries);
-}
-
-/**
- * Loads the leaderboard entries from localStorage.
- * @returns {string[]} - An array of serialized leaderboard entries.
- */
-export function loadLeaderboardEntries(): string[] {
-  const leaderboardEntries = window.localStorage["leaderboard"];
-  if (!leaderboardEntries) {
-    return [];
-  }
-
-  return JSON.parse(leaderboardEntries) || [];
-}
-
-/**
  * Loads the backup state from localStorage.
  * @returns {Promise<BackupState | undefined>} - A promise that resolves to the backup state or undefined if it doesn't exist.
  */
